requirements.txt
natsort==8.4.0
requests==2.32.3
Flask==3.0.3
Flask-Cors==5.0.0
opencv-python==4.11.0.86
numpy==2.2.2
Flask-Session==0.8.0
redis==5.2.1
supabase==2.15.0
python-dotenv==1.1.0